# core-api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Core API for Tremolo (auth, users/teachers/friends, game settings, score entries, classes/assignments, dashboard charts).",
        "title": "Tremolo Core API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/friends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "List friends",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.FriendDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Add a friend",
                "parameters": [
                    {
                        "description": "ID of the user to add",
                        "name": "friend",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.addFriendRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid or missing friend_id, or failed to add",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/friends/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Search users to add as friends",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search text; an empty query returns an empty list",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.FriendDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/note-game/keyboard-bindings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyboard-bindings"
                ],
                "summary": "Get the keyboard bindings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.KeyboardBindingsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No keyboard bindings found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyboard-bindings"
                ],
                "summary": "Update the keyboard bindings",
                "parameters": [
                    {
                        "description": "Bindings to save",
                        "name": "bindings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.KeyboardBindingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.KeyboardBindingsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/note-game/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "note-game-settings"
                ],
                "summary": "Get the note game settings",
                "responses": {
                    "200": {
                        "description": "settings is null if none saved yet",
                        "schema": {
                            "$ref": "#/definitions/dtos.NoteGameSettingsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "note-game-settings"
                ],
                "summary": "Update the note game settings",
                "parameters": [
                    {
                        "description": "Settings to save",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.NoteGameSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.NoteGameSettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or failed to update",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.addFriendRequest": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dtos.FriendDTO": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instrument": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "school": {
                    "type": "string"
                }
            }
        },
...
        "dtos.KeyBindings": {
            "type": "object",
            "properties": {
                "key_a": {
                    "type": "string"
                },
                "key_a_flat": {
                    "type": "string"
                },
                "key_a_sharp": {
                    "type": "string"
                },
                "key_b": {
                    "type": "string"
                },
                "key_b_flat": {
                    "type": "string"
                },
                "key_b_sharp": {
                    "type": "string"
                },
                "key_c": {
                    "type": "string"
                },
                "key_c_flat": {
                    "type": "string"
                },
                "key_c_sharp": {
                    "type": "string"
                },
                "key_d": {
                    "type": "string"
                },
                "key_d_flat": {
                    "type": "string"
                },
                "key_d_sharp": {
                    "type": "string"
                },
                "key_e": {
                    "type": "string"
                },
                "key_e_flat": {
                    "type": "string"
                },
                "key_e_sharp": {
                    "type": "string"
                },
                "key_f": {
                    "type": "string"
                },
                "key_f_flat": {
                    "type": "string"
                },
                "key_f_sharp": {
                    "type": "string"
                },
                "key_g": {
                    "type": "string"
                },
                "key_g_flat": {
                    "type": "string"
                },
                "key_g_sharp": {
                    "type": "string"
                }
            }
        },
        "dtos.KeyboardBindingsRequest": {
            "type": "object",
            "properties": {
                "key_bindings": {
                    "$ref": "#/definitions/dtos.KeyBindings"
                }
            }
        },
        "dtos.KeyboardBindingsResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "key_bindings": {
                    "$ref": "#/definitions/dtos.KeyBindings"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.NoteGameSettingsRequest": {
            "type": "object",
            "properties": {
                "clef": {
                    "type": "string"
                },
                "game_mode": {
                    "type": "string"
                },
                "high_note": {
                    "type": "string"
                },
                "low_note": {
                    "type": "string"
                },
                "note_limit": {
                    "type": "integer"
                },
                "octave": {
                    "type": "integer"
                },
                "scale": {
                    "type": "string"
                },
                "time_limit": {
                    "type": "integer"
                }
            }
        },
        "dtos.NoteGameSettingsResponse": {
            "type": "object",
            "properties": {
                "clef": {
                    "type": "string"
                },
                "game_mode": {
                    "type": "string"
                },
                "high_note": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "low_note": {
                    "type": "string"
                },
                "note_limit": {
                    "type": "integer"
                },
                "octave": {
                    "type": "integer"
                },
                "scale": {
                    "type": "string"
                },
                "time_limit": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}